+++ changelog.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+
+- `molecules/end/EndChip` and `molecules/end/EndChipAndLabel` (internal ref: UR-1612). Test them on the showcase app in `components/ListItems`
+
 ## [0.2.3 - 2020-04-28]
 
 ### Fixed